# backend/api/openapi.yaml
openapi: 3.1.0
...
info:
  title: kubeport API
  version: "0.1.0"
  license:
    name: MIT
    identifier: MIT
  description: |
    kubeport turns Kubernetes resources into a self-service catalog: admins
    publish versioned templates, everyone else deploys them through generated
    forms.

    **This API is a UX layer, not an authorization boundary.** Every write that
    touches a cluster is performed with the caller's own OIDC id_token, forwarded
    unchanged to that cluster's apiserver. Kubernetes RBAC decides what actually
    happens, so a 201 from kubeport only means kubeport accepted the request —
    the cluster may still refuse it.

    ## Two ways in

    The Go API described here listens on `:8080` **inside the cluster only**
    (`ClusterIP`; the Ingress routes all external traffic to the frontend). It
    takes an OIDC id_token as a bearer token.

    Browsers instead talk to the Next.js BFF at `https://<host>/api/v1/*`, which
    authenticates from an httpOnly session cookie and injects the bearer token
    server-side. It **ignores any `Authorization` header you send**. Paths and
    bodies are otherwise identical to this spec.

    So a script port-forwards to the backend Service and sends a bearer
    id_token. Driving the browser session is **not** a client path: the BFF
    ignores `Authorization`, and the session token never leaves the server (it
    is held encrypted in the `sessions` table, not in the cookie).

...
    which, and issue #34 tracks the gap.

...

...
servers:
...
security:
  - bearerAuth: []
...
paths:
...
  /v1/templates/{name}:
    get:
      tags: [templates]
      operationId: getTemplate
      summary: Get one template
      description: |
        A template the list does not show the caller answers 404, as if it did
        not exist: one never published whose drafts the caller may not read
        (kubeport-admin for a global template, a member of the owning team
        otherwise), and on an install with a demo, one on the other side of the
        demo line (#238).
      parameters: [{ $ref: "#/components/parameters/TemplateName" }]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TemplateDetail" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/InternalError" }
    patch:
      tags: [templates]
      operationId: updateTemplate
      summary: Edit template metadata
      description: |
        Only the fields present in the body are changed.

        Every change to an existing template — this route,
        createTemplateVersion, updateTemplateVersion, deleteTemplateVersion,
        publishTemplateVersion, deprecateTemplateVersion and
        undeprecateTemplateVersion — refuses a template the caller cannot see
        (one the template list leaves out for them) with the same 404 as a name
        that matches no template (#244). A 403 with its reason is only for a
        template the caller can see but may not change. Creating a template is
        different: names are unique across the install, so createTemplate's 409
        also collides with a name the caller cannot see.
      parameters: [{ $ref: "#/components/parameters/TemplateName" }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                display_name: { type: string }
                description: { type: string }
                tags:
                  type: array
                  items: { type: string }
      responses:
        "200":
          description: Updated.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Template" }
        "400": { $ref: "#/components/responses/ValidationError" }
        "413": { $ref: "#/components/responses/PayloadTooLarge" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/InternalError" }
...
  /v1/templates/{name}/versions/{v}:
    get:
      tags: [templates]
      operationId: getTemplateVersion
      summary: Get one version
      description: |
        Published and deprecated versions are catalog content that any
        authenticated caller the template list shows them to may read. A draft is
        unreleased authoring state — the whole `resources.yaml`, Secrets
        included — so only the template's owners may read it (kubeport-admin for
        a global template, a member of the owning team otherwise).

        Everything the list hides answers 404, not 403, so the name is not
        confirmed: on an install with a demo, a template on the other side of
        the demo line (#238), and a draft of a template that has never been
        published. Only a draft of an already-published template is refused
        with a 403 that says why.
      parameters:
        - $ref: "#/components/parameters/TemplateName"
        - $ref: "#/components/parameters/VersionNumber"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TemplateVersion" }
        "400": { $ref: "#/components/responses/ValidationError" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/InternalError" }
    patch:
      tags: [templates]
      operationId: updateTemplateVersion
      summary: Edit a draft version
      description: |
        Drafts only. A published or deprecated version is immutable — releases
        are pinned to it.
      parameters:
        - $ref: "#/components/parameters/TemplateName"
        - $ref: "#/components/parameters/VersionNumber"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateVersionRequest" }
      responses:
        "200":
          description: Updated.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TemplateVersion" }
        "400": { $ref: "#/components/responses/ValidationError" }
        "413": { $ref: "#/components/responses/PayloadTooLarge" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409":
          description: The version is not a draft.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Problem" }
        "429": { $ref: "#/components/responses/RateLimited" }
        "500": { $ref: "#/components/responses/InternalError" }
    delete:
      tags: [templates]
      operationId: deleteTemplateVersion
      summary: Delete a draft version
      parameters:
        - $ref: "#/components/parameters/TemplateName"
        - $ref: "#/components/parameters/VersionNumber"
      responses:
        "204": { description: "Deleted. No body." }
        "400": { $ref: "#/components/responses/ValidationError" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409":
          description: The version is not a draft.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Problem" }
        "500": { $ref: "#/components/responses/InternalError" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        An OIDC **id_token** (not an access token) from an issuer this install
        trusts. The backend checks signature, issuer and `aud`, then forwards
        the same token to the target cluster.

        How to obtain one non-interactively is not solved for production IdPs —
        see `docs/machine-clients.md` and issue #34.
  parameters:
...
    TemplateName:
      name: name
      in: path
      required: true
      schema: { type: string }
    VersionNumber:
      name: v
      in: path
      required: true
      schema: { type: integer, minimum: 1 }
...
  headers:
    RequestId:
      description: |
        Echoed on **every** response, and equal to `Problem.request_id` on an
        error. An inbound `X-Request-Id` matching `^[A-Za-z0-9._-]{1,64}$` is
        adopted, so a caller's own id survives the BFF hop; anything else is
        replaced with a generated one.
      schema:
        type: string
        pattern: "^[A-Za-z0-9._-]{1,64}$"
  responses:
    Unauthenticated:
      description: |
        Missing, malformed or rejected token. Through the BFF this also covers a
        missing or expired session cookie.
      headers:
        X-Request-Id: { $ref: "#/components/headers/RequestId" }
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Problem" }
          example:
            type: https://kubeport.io/errors/unauthenticated
            title: unauthenticated
            status: 401
            detail: missing bearer token
            request_id: 4d3f0a2c-7c1e-4a6b-9f10-2f5f1f0b8e21
    Forbidden:
      description: |
        `title` says who refused: `rbac-denied` is kubeport's own check or the
        cluster's RBAC; `demo-restricted` is the demo gate — a management write
        demo accounts never get, template creation or a version's publish,
        deprecate or undeprecate on an
        install without demo.allowTemplateCreate, or a read or change of a release that no demo
        account owns. A template on the other side of the demo line is not a
        403: reading, changing or deploying it answers 404 like a missing one
        (#238, #244).
      headers:
        X-Request-Id: { $ref: "#/components/headers/RequestId" }
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Problem" }
    NotFound:
      description: |
        No such resource. On template routes this is also the answer for a
        template the caller cannot see — one the template list leaves out for
        them — whether they read, render or change it, or deploy one across
        the demo line (#226, #238, #244); a 403 on those routes means the
        caller can see the template.
      headers:
        X-Request-Id: { $ref: "#/components/headers/RequestId" }
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Problem" }
    ValidationError:
      description: The request body or a parameter did not validate.
      headers:
        X-Request-Id: { $ref: "#/components/headers/RequestId" }
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Problem" }
    PayloadTooLarge:
      description: |
        The request body is over the 4 MiB cap (`payload-too-large`, #128).
        Until #128 this was a `400 validation-error` whose `detail` was Go's
        `http: request body too large`; it has its own kind because the
        response is the opposite one — a validation error is fixed by
        correcting the body, this one only by sending less of it. **Do not
        retry the same request.** Shrink it, or split a template into several.

        A `Content-Length` over the cap is refused before routing and
        authentication, on every operation and method — so an unauthenticated
        caller sees 413 rather than 401, and a GET that declares such a length
        is refused even though no GET reads a body. A body of unknown length
        (chunked) is cut off at the cap while the handler reads it, with the
        same response. Only operations that take a request body list this
        response, since those are the ones a well-formed client can reach it on.

        The server does not drain the rest of an oversized body and closes the
        connection after answering, so a client still uploading may see the
        connection drop before it reads this. Measure before sending.
      headers:
        X-Request-Id: { $ref: "#/components/headers/RequestId" }
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Problem" }
          example:
            type: https://kubeport.io/errors/payload-too-large
            title: payload-too-large
            status: 413
            detail: request body exceeds 4 MiB
            request_id: 4d3f0a2c-7c1e-4a6b-9f10-2f5f1f0b8e21
    RateLimited:
      description: |
        Over the per-caller budget (`rate-limited`). The budget is a token
        bucket keyed by OIDC subject, in-process and per replica — it exists to
        bound how often a caller can make kubeport talk to a cluster's control
        plane, not to price the API (issue #73).

        `Retry-After` says how many seconds to wait; honour it rather than
        retrying at once or sleeping an arbitrary constant. The demo's two Dex
        accounts are static, so concurrent demo visitors share one bucket.

        Budgets are per route group, not one per caller: reading a release
        (`GET /v1/releases/{id}`, 240/min) does not spend the budget shared by
        the SSAR, cluster-openapi and log-stream routes (60/min); creating,
        updating and deleting a release share one of their own (30/min), so
        switching verbs does not reset it; and the template authoring and
        preview routes have their own. The route decides
        the budget; `X-RateLimit-Limit` repeats its size, and two budgets are
        both 60/min. The `X-RateLimit-*` pair is sent only on a 429 — a
        successful response carries no remaining count, so a polling client
        paces itself. On the demo, every visitor of one Dex account spends the
        same 240/min of release reads.

        `too-many-streams` is the other 429, and only
        `GET /v1/releases/{id}/logs` gives it: the caller already holds as many
        log streams **open** as it may (#169). The bucket above prices opening a
        stream and nothing after, so it cannot bound how many a caller keeps —
        and each keeps a goroutine per pod and an apiserver connection alive.
        Branch on `title`: this one clears when one of the caller's streams
        closes, not after a set time, so close a stream you no longer need
        rather than only backing off. `Retry-After` is still present, as a
        backoff hint rather than a prediction, and the `X-RateLimit-*` pair is
        **not** — it describes the per-minute budget and would be wrong about a
        count of open streams. The cap is per caller, so the demo's shared
        accounts share it too — and each sign-in to a demo account also has a
        cap of 8 of its own (#200), so one sign-in cannot hold the whole pool.
        Signing in again brings a new cap, so this raises what holding the pool
        costs rather than preventing it.
      headers:
        Retry-After:
          description: Seconds to wait before retrying. Always present on a 429.
          schema: { type: integer, minimum: 1 }
        X-RateLimit-Limit:
          description: Requests allowed per minute for this caller.
          schema: { type: integer }
        X-RateLimit-Remaining:
          description: Always `0` here — the header is emitted only on refusal.
          schema: { type: integer }
        X-Request-Id: { $ref: "#/components/headers/RequestId" }
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Problem" }
...
    InternalError:
      description: |
        Something failed on kubeport's side. `title` is `internal` for a
        database, rendering or k8s-client-construction failure, or
        `cluster-config` when the stored registration's TLS material is not
        usable.

        **`detail` no longer carries the underlying error** (issue #49). A pgx
        connect failure spelled out `host=… user=… database=…`, and a k8s
        transport failure carried the apiserver's address — the same value
        `GET /v1/clusters` deliberately withholds (#52). The reason is in the
        server log; quote `request_id` to have it looked up.
      headers:
        X-Request-Id: { $ref: "#/components/headers/RequestId" }
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Problem" }
  schemas:
    Problem:
      type: object
      description: |
        RFC 7807-shaped. Every error this API and the BFF *route* uses it —
        branch on `title`, the stable machine-readable kind; `detail` is prose
        and may change.

        Two caveats worth knowing before you write a parser:

        - It is served as `application/json`, not `application/problem+json`
          (`errors.go` uses gin's `AbortWithStatusJSON`). A strict RFC 7807
          client that content-negotiates on the media type will not match.
        - There is no non-JSON error left to handle. A path matching no route
          is `404 not-found`, a wrong verb is `405 method-not-allowed`, and a
          panic is `500 internal` — all in this schema. Until #81 the first two
          were gin's `text/plain` defaults and the third had no body at all, so
          a parser written before that needed a content-type fallback. It does
          not any more. See "Every error is a Problem" above.
      required: [type, title, status]
      properties:
        type:
          type: string
          format: uri
          description: "`https://kubeport.io/errors/<title>`. Not dereferenceable."
        title: { $ref: "#/components/schemas/ErrorKind" }
        status: { type: integer }
        detail: { type: string }
        demo_policy:
          type: array
          description: |
            Only on `403 demo-restricted` from `POST /v1/releases`,
            `PUT /v1/releases/{id}` and `POST /v1/templates/{name}/render` (#350):
            where a demo account's rendered manifest breaks the demo's limits.
            Every violation is listed at once. Accounts outside the demo, and
            installs without demo mode, never get it.

            The limits bound what a visitor leaves on the node. A limit the
            template leaves unset is filled in rather than refused — the 200
            and `rendered_yaml` show the filled manifest — so this lists only
            values set above a limit, and images outside the allowed prefixes.
            The allowed prefixes themselves are never returned.
          items:
            type: object
            required: [rule, kind, name, field]
            properties:
              rule:
                type: string
                enum: [job-backoff-limit, job-ttl, cronjob-history-limit, image-prefix]
                description: |
                  `job-backoff-limit`: a Job's, or a CronJob job template's,
                  `backoffLimit`. `job-ttl`: a Job's own `ttlSecondsAfterFinished`
                  (a CronJob's jobs are bounded by its history instead).
                  `cronjob-history-limit`: `successfulJobsHistoryLimit` or
                  `failedJobsHistoryLimit`. `image-prefix`: a container or init
                  container image the demo does not allow.
              kind: { type: string }
              name: { type: string }
              container:
                type: string
                description: The container, for `image-prefix`.
              field:
                type: string
                description: The path within the object, e.g. `spec.jobTemplate.spec.backoffLimit`.
              limit:
                type: integer
                description: The cap, for the numeric rules.
              got:
                description: The value the manifest holds.
        conflicts:
          type: array
          description: |
            Only on `409 resource-conflict` (#161): each object the release
            would have taken over, and what holds it. Name the holder from
            here, not by parsing `detail`.

            Objects are checked with the caller's token. One the caller may
            write but not read is still found on a create, because a server-side
            dry run shows it exists, and appears with `owner_unknown`. On an
            update it cannot be told apart from the release's own object, so it
            is not checked, and a 200 does not prove no such object was taken
            over.
          items:
            type: object
            required: [kind, name, namespace, owner]
            properties:
              kind: { type: string }
              name: { type: string }
              namespace: { type: string }
              owner:
                type: string
                description: |
                  The release holding the object, or "" when kubeport did not
                  create it or when `owner_unknown` is set.
              owner_unknown:
                type: boolean
                description: |
                  The object exists but the caller may not read it, so its
                  holder could not be looked up. Omitted when false.
              same_name:
                type: boolean
                description: |
                  `owner` is the name of the release being deployed, but the
                  object carries another release's id (#195): it was left by an
                  earlier release of that name, or is held by a release of that
                  name under another registration of the same cluster. Omitted
                  when false.
        pinned_namespace:
          type: object
          description: |
            Only on a `400 validation-error` caused by a template object that
            pins a `metadata.namespace` other than the release's (#137). The
            template is what has to change, not the request, so a client can
            say so instead of asking the user to check their input.
          required: [kind, name, namespace]
          properties:
            kind: { type: string }
            name: { type: string }
            namespace:
              type: string
              description: The namespace the template pins.
        template_defect:
          type: object
          description: |
            Only on a `400 validation-error` from creating or updating a release
            whose template version has a ui-spec field of a type kubeport does
            not know (#136). Saving such a spec is refused now, but a version
            saved before that still exists, and no request values deploy it —
            the template has to change, as with `pinned_namespace`.
          required: [path, type]
          properties:
            path:
              type: string
              description: The ui-spec path of the field.
            type:
              type: string
              description: The field type as stored.
        request_id:
          type: string
          description: |
            Present on every 4xx and 5xx, and equal to the `X-Request-Id`
            response header. Quote it when reporting a failure: a 500 keeps its
            reason in the server log rather than the body, and this is what ties
            the two together.

            An inbound `X-Request-Id` is adopted when it matches
            `^[A-Za-z0-9._-]{1,64}$`, so a caller's own trace survives the BFF
            hop; anything else is replaced. The charset is enforced because
            header values may contain spaces and `=`, which would let a crafted
            id forge fields inside the access-log line.
...
    Template:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        display_name: { type: string }
        description: { type: [string, "null"] }
        tags:
          type: array
          items: { type: string }
        owner_user_id: { type: string, format: uuid }
        current_version_id: { type: [string, "null"], format: uuid }
        owning_team_id: { type: [string, "null"], format: uuid }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    TemplateDetail:
      allOf:
        - $ref: "#/components/schemas/Template"
        - type: object
          properties:
            current_version: { type: [integer, "null"] }
            owning_team_name: { type: [string, "null"] }
...
    TemplateVersion:
      type: object
      properties:
        id: { type: string, format: uuid }
        template_id: { type: string, format: uuid }
        version: { type: integer }
        resources_yaml: { type: string }
        ui_spec_yaml:
          type: string
          contentMediaType: application/yaml
          contentSchema: { $ref: "#/components/schemas/UiSpec" }
        metadata_yaml: { type: [string, "null"] }
        status:
          type: string
          enum: [draft, published, deprecated]
        notes: { type: [string, "null"] }
        authoring_mode:
          type: string
          enum: [yaml, ui]
        ui_state_json:
          type: [object, "null"]
          additionalProperties: true
        created_by_user_id: { type: string, format: uuid }
        created_at: { type: string, format: date-time }
        published_at: { type: [string, "null"], format: date-time }
...
    UpdateVersionRequest:
      type: object
      description: Only the fields present are changed.
      properties:
        resources_yaml: { type: string }
        ui_spec_yaml:
          type: string
          contentMediaType: application/yaml
          contentSchema: { $ref: "#/components/schemas/UiSpec" }
        ui_state:
          type: object
          additionalProperties: true
        metadata_yaml: { type: string }
        notes: { type: string }
...